refactor(macros): split DIContainerCodeGenerator into focused files
DIContainerCodeGenerator.swift grew to 776 LOC as we piled in factory,
dependency-resolution, and AST helpers. Break it into four files along
natural concerns so each piece is reviewable in isolation:

  - DIContainerCodeGenerator.swift (398 LOC): struct entry point +
    primary init builder + AsyncTaskBinding.
  - DIContainerCodeGenerator+Factory.swift (131 LOC): makeFactoryExpr,
    makeAsyncFactoryExpr, labeledDependencyArguments.
  - DIContainerCodeGenerator+Dependency.swift (206 LOC): closure parameter
    resolution and _innoDIUnresolvedDependency fallback helpers.
  - DIContainerCodeGenerator+Helpers.swift (94 LOC): small AST utilities
    (access modifiers, main-actor attribute, assignExpr, etc.).

Drive-by: remove `mapDependencyNameToStorageName` — it had no callers.
Emitted syntax is byte-identical; macro snapshots unchanged.
